requirements.txt
# QA Automation Scripts - Python Dependencies

# Core PDF processing libraries
PyMuPDF>=1.23.0
pdfplumber>=0.9.0
PyPDF2>=3.0.1

# Document processing
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.1.2

# Markdown and HTML processing
markdown>=3.5.1
beautifulsoup4>=4.12.2

# Text processing and similarity
scikit-learn>=1.3.0
difflib2>=0.1
langdetect>=1.0.9

# Web scraping and HTTP requests (for API testing)
requests>=2.31.0

# Data analysis and visualization
pandas>=2.1.0

# Image processing
Pillow>=10.0.0

# Testing framework
pytest>=7.4.0
pytest-html>=3.2.0

# Optional: Additional libraries for enhanced functionality
# lxml>=4.9.3          # For XML processing
# nltk>=3.8.1          # For advanced text processing
# spacy>=3.7.0         # For NLP tasks
# textdistance>=4.6.0  # For advanced text similarity metrics